Moved dash::Halo back to regular include path
